# api_specs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Currency Conversion API",
    "description": "A demo API for currency conversion with debugging capabilities",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/rates/history": {
      "get": {
        "tags": [
          "rates"
        ],
        "summary": "Get Rates History",
        "description": "Get historical exchange rates with optional filtering.\n\nArgs:\n    currency: Optional currency code to filter results\n    days: Number of days of history to retrieve (1-365)\n    limit: Maximum number of records to return (1-10000)\n    db: Database session\n\nReturns:\n    Historical exchange rates data\n\nRaises:\n    HTTPException: If history cannot be retrieved or parameters are invalid",
        "operationId": "get_rates_history_api_v1_rates_history_get",
        "parameters": [
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by specific currency code",
              "title": "Currency"
            },
            "description": "Filter by specific currency code"
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "description": "Number of days of history to retrieve",
              "default": 7,
              "title": "Days"
            },
            "description": "Number of days of history to retrieve"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 1,
              "description": "Maximum number of records",
              "default": 1000,
              "title": "Limit"
            },
            "description": "Maximum number of records"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatesHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/debug/user-activity": {
      "get": {
        "tags": [
          "debug"
        ],
        "summary": "Get User Activity",
        "description": "Get recent user activity for DDoS investigation.\n\nThis endpoint helps identify users with unusually high activity levels\nby analyzing conversion requests over the specified time window.\n\nArgs:\n    minutes: Number of minutes to look back (1-60)\n    limit: Maximum number of users to return (5-100)\n    db: Database session\n\nReturns:\n    Dictionary with user activity statistics",
        "operationId": "get_user_activity_api_debug_user_activity_get",
        "parameters": [
          {
            "name": "minutes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 60,
              "minimum": 1,
              "description": "Minutes to look back",
              "default": 10,
              "title": "Minutes"
            },
            "description": "Minutes to look back"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 5,
              "description": "Max users to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Max users to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get User Activity Api Debug User Activity Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/debug/request-patterns": {
      "get": {
        "tags": [
          "debug"
        ],
        "summary": "Get Request Patterns",
        "description": "Analyze request patterns for anomaly detection.\n\nThis endpoint provides insights into request patterns that might indicate\ncoordinated attacks or unusual behavior.\n\nArgs:\n    minutes: Number of minutes to look back (1-60)\n    db: Database session\n\nReturns:\n    Dictionary with request pattern analysis",
        "operationId": "get_request_patterns_api_debug_request_patterns_get",
        "parameters": [
          {
            "name": "minutes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 60,
              "minimum": 1,
              "description": "Minutes to look back",
              "default": 10,
              "title": "Minutes"
            },
            "description": "Minutes to look back"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Request Patterns Api Debug Request Patterns Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoricalRateInfo": {
        "properties": {
          "currency": {
            "type": "string",
            "title": "Currency",
            "description": "Currency code"
          },
          "rate": {
            "type": "string",
            "title": "Rate",
            "description": "Exchange rate relative to base currency"
          },
          "recorded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Recorded At",
            "description": "When this rate was recorded"
          },
          "base_currency": {
            "type": "string",
            "title": "Base Currency",
            "description": "Base currency for this rate",
            "default": "USD"
          },
          "rate_source": {
            "type": "string",
            "title": "Rate Source",
            "description": "Source of the rate data",
            "default": "simulated"
          }
        },
        "type": "object",
        "required": [
          "currency",
          "rate",
          "recorded_at"
        ],
        "title": "HistoricalRateInfo",
        "description": "Historical exchange rate information with timestamp."
      },
...
      "RatesHistoryResponse": {
        "properties": {
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency",
            "description": "Specific currency (if filtered)"
          },
          "rates": {
            "items": {
              "$ref": "#/components/schemas/HistoricalRateInfo"
            },
            "type": "array",
            "title": "Rates",
            "description": "List of historical rates"
          },
          "period": {
            "additionalProperties": {
              "type": "string",
              "format": "date-time"
            },
            "type": "object",
            "title": "Period",
            "description": "Time period covered"
          },
          "total_records": {
            "type": "integer",
            "title": "Total Records",
            "description": "Total number of records returned"
          },
          "base_currency": {
            "type": "string",
            "title": "Base Currency",
            "description": "Base currency for all rates",
            "default": "USD"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "When the history was retrieved"
          },
          "metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Metadata",
            "description": "Additional metadata about the historical data"
          }
        },
        "type": "object",
        "required": [
          "rates",
          "period",
          "total_records"
        ],
        "title": "RatesHistoryResponse",
        "description": "Response model for historical exchange rates."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}